refactor: move Tiingo provider to provider/tiingo sub-package
