+++ cmd/providers_register.go
@@ -19,4 +19,5 @@ import (
 	_ "github.com/penny-vault/pvdata/provider/fred"
 	_ "github.com/penny-vault/pvdata/provider/legacy"
 	_ "github.com/penny-vault/pvdata/provider/massive"
+	_ "github.com/penny-vault/pvdata/provider/tiingo"
 )
